+++ Topos.Kafka.Tests/KafkaCloudTest.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Concurrent;
+using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using NUnit.Framework;
@@ -7,61 +8,66 @@ using Testy;
 using Testy.Extensions;
 using Topos.Config;
 using Topos.Producer;
+using Topos.Serialization;
 // ReSharper disable ArgumentsStyleAnonymousFunction
 #pragma warning disable 1998
-
 // ReSharper disable ArgumentsStyleOther
 // ReSharper disable ArgumentsStyleStringLiteral
 
 namespace Topos.Kafka.Tests
 {
+    [Explicit]
     [TestFixture]
     public class KafkaCloudTest : FixtureBase
     {
-        const string Key = "paste-key-here";
-        const string Secret = "paster-secret-here";
+        string host;
+        string key;
+        string secret;
+
+        protected override void SetUp()
+        {
+            var lines = File.ReadAllLines(Path.Combine(AppContext.BaseDirectory, "confluent_cloud.secret.txt"));
+
+            host = lines[0];
+            key = lines[1];
+            secret = lines[2];
+        }
 
         [Test]
-        [Explicit]
         public async Task CanConnectToKafkaCloud()
         {
             const string topic = "test-topic";
 
             using var producer = Configure
-                .Producer(p =>
-                {
-                    p.UseKafka("pkc-lz6r3.northeurope.azure.confluent.cloud:9092")
-                        .WithConfluentCloud(Key, Secret);
-                })
+                .Producer(p => p.UseKafka(host).WithConfluentCloud(key, secret))
                 .Serialization(s => s.UseNewtonsoftJson())
                 .Create();
 
             await producer.Send(topic, new ToposMessage("her er en tekststreng"));
 
-            var receivedStrings = new ConcurrentQueue<string>();
+            var receivedMessages = new ConcurrentQueue<ReceivedLogicalMessage>();
 
             using var consumer = Configure
-                .Consumer("default", c =>
-                {
-                    c.UseKafka("pkc-lz6r3.northeurope.azure.confluent.cloud:9092")
-                        .WithConfluentCloud(Key, Secret);
-                })
+                .Consumer("default", c => c.UseKafka(host).WithConfluentCloud(key, secret))
                 .Serialization(s => s.UseNewtonsoftJson())
                 .Topics(t => t.Subscribe(topic))
                 .Positions(p => p.StoreInFileSystem(NewTempDirectory()))
                 .Handle(async (events, context, token) =>
                 {
-                    events.DumpJson();
-                    foreach (var str in events.Select(e => e.Body).OfType<string>())
+                    foreach (var evt in events)
                     {
-                        receivedStrings.Enqueue(str);
+                        receivedMessages.Enqueue(evt);
                     }
                 })
                 .Create();
 
             consumer.Start();
 
-            await Task.Delay(TimeSpan.FromMinutes(1));
+            await receivedMessages.WaitOrDie(q => q.Count >= 1, timeoutSeconds: 20);
+
+            await Task.Delay(TimeSpan.FromSeconds(1));
+
+            receivedMessages.Select(s => new { Pos = s.Position, Msg = s.Body?.ToString() }).DumpTable();
         }
     }
 }