+++ CHANGELOG.md
@@ -44,3 +44,6 @@
 ## 0.0.46
 * Update Kafka driver to 1.3.0
 * Add configuration helper for connecting to Confluent Cloud
+
+## 0.0.47
+* React quicker on closed connections in consumer